crud en java para la entidad producto

+++ stockflow-database/schema.sql
@@ -31,7 +31,7 @@ CREATE TABLE products (
     product_code SERIAL PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
     description TEXT,
-    price NUMERIC(10,2) NOT NULL CHECK (price >= 0),
+    price INTEGER NOT NULL CHECK (price >= 0),
     active BOOLEAN DEFAULT true,
     created_at TIMESTAMP WITHOUT TIME ZONE DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP WITHOUT TIME ZONE DEFAULT CURRENT_TIMESTAMP
